+++ fltk/src/tree.rs
@@ -1149,6 +1149,23 @@ impl Tree {
         assert!(!self.was_deleted());
         unsafe { mem::transmute(Fl_Tree_callback_reason(self.inner)) }
     }
+
+    /// Get an item's pathname
+    pub fn item_pathname(&self, item: &TreeItem) -> Result<String, FltkError> {
+        assert!(!self.was_deleted());
+        let mut temp = vec![0u8; 256];
+        unsafe {
+            let ret = Fl_Tree_item_pathname(self.inner, temp.as_mut_ptr() as _, 256, item.inner);
+            if ret == 0 {
+                if let Some(pos) =  temp.iter().position(|x| *x == 0) {
+                    temp = temp.split_at(pos).0.to_vec();
+                }
+                Ok(String::from_utf8_lossy(&temp).to_string())
+            } else {
+                Err(FltkError::Internal(FltkErrorKind::FailedOperation))
+            }
+        }
+    }
 }
 
 impl IntoIterator for Tree {
@@ -1835,6 +1852,11 @@ impl TreeItem {
         } else {
             unsafe { Fl_Tree_Item_set_usericon(self.inner, std::ptr::null_mut::<raw::c_void>()) }
         }
+    }
+
+    /// Return the internal pointer of the tree item
+    pub fn as_ptr(&self) -> *mut Fl_Tree_Item {
+        self.inner
     }
 }