+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 ## [1.2.19] - Unreleased
 - Fix docs for dialog::message_title_default(). Thanks @hannesbraun.
 - Add TreeItem::set_label_fgcolor() and label_fgcolor(), deprecate the older names. Thanks @AshfordN.
+- Add TreeItem::as_ptr().
+- Add Tree::item_pathname().
 
 ## [1.2.18] - 2021-11-27
 - Add menu::mac_set_about(). Thanks @hannesbraun.